fix(unittests): move keystore config tests to unit_tests/test_config.py
branch-perf-v17 predates the move of config tests to unit_tests/unit/,
and the backported copy of master's unit_tests/unit/test_config.py
imports sdcm.provision.aws.capacity_errors, which does not exist on
this branch, breaking test collection. Drop the stray file and apply
the keystore test additions from the series to the branch's
unit_tests/test_config.py instead, matching the 2026.1/2025.1
backports.
